[`refined4s-circe`] Update test structure
